feat(frontend): Refactor App.tsx to minimal shell with extracted HomePage
## PR7: App Shell Refactoring

### Changes Made:
- Refactored App.tsx to be a minimal shell component
- Created AppShell component to handle routing logic
- Extracted HomePage as a separate page component with CSS Modules
- Created navigation feature module with:
  - TabNavigation component with CSS Modules
  - useNavigation custom hook for URL synchronization
  - Type definitions for navigation
- Centralized tab configuration in config/tabs.config.ts
- Moved all HomePage styles to HomePage.module.css

### Architecture Improvements:
- Clean separation of concerns with minimal App.tsx
- Feature-based navigation module
- Component-scoped CSS with CSS Modules
- Reusable navigation logic via custom hook
- Centralized configuration for tabs

### Tests:
- All 177 tests passing
- ESLint clean with no errors
